Fixed a few formatting errors and updated
Added immutable v mutable, but unsure where to put it so I'll leave that up to you. Fixed some formatting errors where the '```' needs to be put at the beginning of the line for it to take effect properly otherwise the blocks say objective-c at the start and dont have word colouring etc.